Suddivise le classi in GUI, BusinnesLogic, DB
