+++ src/code/Businnes_Logic/CashDispenser.java
@@ -1,4 +1,4 @@
-package code;
+package code.Businnes_Logic;
 // CashDispenser.java
 // Represents the cash dispenser of the ATM
