+++ src/code/Businnes_Logic/DepositSlot.java
@@ -1,24 +1,16 @@
-package code;
-// Keypad.java
-// Represents the keypad of the ATM
-import java.util.Scanner; // program uses Scanner to obtain user input
+package code.Businnes_Logic;
+// DepositSlot.java
+// Represents the deposit slot of the ATM
 
-public class Keypad
+public class DepositSlot 
 {
-   private Scanner input; // reads data from the command line
-                         
-   // no-argument constructor initializes the Scanner
-   public Keypad()
+   // indicates whether envelope was received (always returns true, 
+   // because this is only a software simulation of a real deposit slot)
+   public boolean isEnvelopeReceived()
    {
-      input = new Scanner( System.in );    
-   } // end no-argument Keypad constructor
-
-   // return an integer value entered by user 
-   public int getInput()
-   {
-      return input.nextInt(); // we assume that user enters an integer  
-   } // end method getInput
-} // end class Keypad  
+      return true; // deposit envelope was received
+   } // end method isEnvelopeReceived
+} // end class DepositSlot
