+++ src/code/Businnes_Logic/Transaction.java
@@ -1,6 +1,9 @@
-package code;
+package code.Businnes_Logic;
 // Transaction.java
 // Abstract superclass Transaction represents an ATM transaction
+
+import code.Database.BankDatabase;
+import code.GUI.Screen;
 
 public abstract class Transaction
 {
