+++ src/code/GUI/BalanceInquiry.java
@@ -1,6 +1,9 @@
-package code;
+package code.GUI;
 // BalanceInquiry.java
 // Represents a balance inquiry ATM transaction
+
+import code.Businnes_Logic.Transaction;
+import code.Database.BankDatabase;
 
 public class BalanceInquiry extends Transaction
 {
